chore: remove temporary reviewdog fix (#12284)
 - [x] depends on: https://github.com/reviewdog/action-suggester/pull/52


This reverts #12280 after the upstream fix
 https://github.com/reviewdog/reviewdog/issues/1696
has propagated to reviewdog/action-suggester@v1. This should happen when 
https://github.com/reviewdog/action-suggester/releases
has a release that uses [v0.17.4](https://github.com/reviewdog/reviewdog/releases/tag/v0.17.4) or higher.

When this happens, rebase this pull request and re-trigger to see that it doesn't fail with the 300 dummy files. Then remove the dummy files and merge this pull request.



Co-authored-by: Moritz Firsching <firsching@google.com>