refactor: Redefine 3AP-free sets (#12701)
The current definition allows `{0, 1}` as a 3AP-free set in `ZMod 2`, which is really inconvenient from the point of view of interfacing with corners. This PR redefines it so that  `0, 1, 0` in `ZMod 2` is a 3AP and `{0, 1}` is not 3AP-free.

Also take the opportunity to reorder the binders to a more convenient fashion, to move the heavy result `threeAPFree_sphere` to `Combinatorics.Additive.Behrend` where it is used (it won't be used anywhere else) and to drop `rothNumberNat_isBigO_id` since it's a trivial result that pulls in a bunch of analysis.