refactor: Turn Freiman homs into predicates (#12546)
The bundled hom approach to Freiman homomorphisms is a fail. It is very hard to use and doesn't bring anything. That is because Freiman homs don't have much structure and we do not study them under this angle.

This PR replaces the bundled homs by predicates and moves the file from `Algebra.Group.Freiman` to `Combinatorics.Additive.FreimanHom` so that:
* It is clear these are combinatorial objects, not algebraic ones.
* The diff isn't completely mangled. Nothing was kept so it's not worth comparing via a naïve git diff.

Also fix a few oddities accidentally introduced in #12701.