chore(CategoryTheory): fix hasFiniteLimits_of_hasLimitsOfSize (#12797)
Given a category `C`, `hasFiniteLimits_of_hasLimitsOfSize` (and the dual statement for colimits) involves two universes `u'` and `v'` which may be unrelated to the two universes `u` and `v` involved in the category `C : Type u` (and `[Category.{v} C]`). For general universes, it is made a lemma instead of an instance. It is made an instance only in the `_hasLimits` case (which corresponds to `u'=v'=v`) and the `u'=v'=0` case.

This fixes the behavior observed in https://leanprover.zulipchat.com/#narrow/stream/113489-new-members/topic/invalid.20occurrence.20of.20universe.20level.20'u_1'/near/437497392