feat: the homological functor on the homotopy category of cochain complexes in an abelian category (#11760)
In this PR, it is shown that the homology functors on the homotopy category of cochain complexes in an abelian category are homological functors (in the sense that they send distinguished triangles to exact sequences). This PR also shows that the category of homological complexes in an abelian category is an abelian category.