CI: add a step producing a "reduced diff" (#12515)
Add an extra CI step whose purpose is to produce a summary of declarations that have been added, removed or simply moved around unchanged.

This can help review "moving" PRs, whose sole purpose is to move code around.

The script is text-based, so that it can run right away and does not have to wait for the output of `lake build`.  As a consequence, it is easy to fool, but also gives out answers quickly!