chore: Move basic `Int` lemmas earlier A bunch of lemmas about `Int` can be moved to `Data.Int.Defs` at little cost. This PR moves them and deletes the now empty `Data.Int.Dvd.Basic`.
The story is that those lemmas used to rely on lemmas proved for algebraic order structures, hence required the algebraic order instances on `Int` to be available. In Lean 4, those preliminary lemmas have been special-cased to `Nat` and `Int`, so there is no need for the general instances anymore (at least in basic files).