feat(CategoryTheory/Galois): prorepresentability of fiber functors (#12839)
We show that a fiber functor of a Galois category is pro-represented by the system of pointed Galois objects. We also introduce the system of automorphism groups (seen as objects of `C`) of the pointed Galois objects.

This is used in a follow-up PR to establish a group isomorphism between the automorphism group of a fiber functor and the limit of the automorphism groups of the Galois objects.