chore: adaptations for nightly-2024-05-14 (#12902)
The main underlying change here is leanprover/lean4#4119, which fixed a
number of typeclass search bugs. In particular, it fixed a caching bug
(the `synthPendingDepth` was being ignored during caching, resulting in
both false positive and false negatives). For the false positives (i.e.
things that synthesis was finding that it really shouldn't have been!),
we've had to introduce `set_option maxSynthPendingDepth 2 in` in places.
These are all explained with `#adaptation_notes`.

There's also leanprover/lean4#4061, which makes well-founded definitions
irreducible by default. This helps avoid some horrible run-away
unfoldings, and was surprisingly un-painful to Mathlib to (credit to
@nomeata for discovering this!) There are some minor adaptations still
required, however.

---------

Co-authored-by: sgouezel <sebastien.gouezel@univ-rennes1.fr>