chore: Remove algebra imports to `Data.Fin.Basic` (#12862)
Move the algebraic instance on `Fin n` from `Data.Fin.Basic` to a new file `Algebra.Group.Fin`. I credit Yakov for https://github.com/leanprover-community/mathlib/pull/5010.

The one bit of algebra that is still imported in `Data.Fin.Basic` is `NeZero`. This is in prevision of its future upstreaming to Batteries or Std.