feat(NumberTheory/EllipticDivisibilitySequence): expose the auxiliary sequence (#10814)
Expose the auxiliary sequences associated to normalised elliptic divisibility sequences along with their corresponding API, and fix the comments under implementation notes. This is needed to define $n$-division polynomials of elliptic curves in a later downstream file as univariate polynomials with the factor of the bivariate $2$-division polynomial omitted.