feat: the equational criterion for vanishing (#12647)
Let $M$ and $N$ be modules over a commutative ring $R$. Following Lemma 8.16 of Altman and Kleiman's *A term of commutative algebra*, we prove some results about under what circumstances a sum of pure tensors $\sum_i m_i \otimes n_i \in M \otimes_R N$ vanishes. The *equational criterion for flatness* is not included in this pull request, but it is an easy consequence of these results.