chore: backports anticipating nightly-2024-05-19 (#13046)
This backports some changes that fill in a few implicit arguments.

nightly-2024-05-19 has trouble with these and the bug fix in https://github.com/leanprover/lean4/pull/4206, and needs some fixes.

I *think* all these changes are not actually problematic, and perhaps even improve readability.

If a maintainer agrees, let's just do this on master. If any of these look at all bad, please just say so and I'll remove it from this PR and consider doing a minimization to understand if/why it is really necessary.