chore (CategoryTheory): remove references to `restate_axiom` and porting notes about `reassoc.1` (#13070)
As noted by @semorrison early in the port:  `restate_axiom` should not be necessary in lean4. Hopefully we can just remove the backticks from field names, then delete the invocation of `restate_axiom`. We delete these references to `restate_axiom` in this PR. We also delete porting notes which just note that `reassoc` syntax is different in lean4.