feat: Fourier transform of iterated derivatives (#12769)
We show that the Fourier transform of an iterated derivative is the Fourier transform of the original function multiplied by a power function. We deduce bounds on `‖w‖ ^ n * ‖iteratedFDeriv ℝ k (𝓕 f) w‖` in terms of integrals of iterated derivatives of `f` (up to order `n`) multiplied by power functions (up to order `k`).