feat: Fourier transform as a continuous linear equiv on Schwartz space (#12144)
We define the Fourier transform on the Schwartz space, as a continuous linear equivalence. First, we define it as a continuous linear map (which amounts to showing that the Fourier transform of a Schwartz function is smooth, with explicit bounds for the successive derivatives in terms of various Schwartz seminorms). Then, we extend it to a continuous linear equivalence thanks to Fourier inversion.