feat(CategoryTheory): surjective-injective factorizations in concrete categories (#13386)
In this PR, we formalize the notion of surjective/injective factorizations in concrete categories. We show that this obviously holds in the category of types. In future PRs, it shall be promoted to locally-surjective/locally-injective factorizations in categories of (pre)sheaves, and this shall be used in order to relate locally surjective morphisms and epimorphisms in categories of sheaves (beyond the case of sheaves of types which was obtained in #7721 and #13318).