feat(CategoryTheory/Sites): locally bijective morphisms of presheaves (#13324)
Let `C` a category equipped with a Grothendieck topology `J`. Let `A` be a concrete category. In this file, we introduce a type-class `J.WEqualsLocallyBijective A` which says that the class `J.W` (of morphisms of presheaves which become isomorphisms after sheafification) is the class of morphisms that are both locally injective and locally surjective (i.e. locally bijective). We prove that this holds iff for any presheaf `P : Cᵒᵖ ⥤ A`, the sheafification map `toSheafify J P` is locally bijective. We show that this holds under certain universe assumptions.



Co-authored-by: Joël Riou <37772949+joelriou@users.noreply.github.com>