chore: clean up `nolints.yml` workflow (#13351)
Synchronise the early steps with the main build workflow: the differences were minor and not essential, but there is little point in being different. Best reviewed by `diff`ing against `build.yml`.

It would be nice to avoid the copy-pasting of e.g. the build mathlib jobs: that pre-existing issue is left for future PRs.