chore: remove unnecessary outParam (#13067)
I spotted an `outParam` in an instance argument of a class. The purpose of `outParam` is to fill in the value during type class search. However, since this is an instance argument, it must already have been synthesised at this point, so this has no purpose.

I looked at the old Zulip conversation on this topic, and there didn't seem to be a good reason for putting this `outParam` in the instance argument, other than simply that it worked.