feat(FiberedCategory/HomLift): definition of IsHomLift (#12978)
We define the notion `IsHomLift` which aims to provide API for working with equalities `p(\phi) = f` for a functor `p` and morphisms `\phi`, `f`. This API is extensively used in an upcoming PR defining the notion of fibered categories. See the repository [Stacks-project](https://github.com/Paul-Lez/Stacks-project) for WIP using this notion.

Co-authored-by: Paul Lezau <paul.lezeau@gmail.com>