feat: Levy-Prokhorov distance pseudometrizes convergence in distribution (#11549)
Add the more interesting direction of topology comparison for the Lévy-Prokhorov distance. Conclude that if the underlying space is separable, then the topology of convergence in distribution coincides with the one obtained from the Lévy-Prokhorov (pseudo)metric. In particular the topology of convergence in distribution is (pseudo)metrizable, and so first countable.



Co-authored-by: kkytola <39528102+kkytola@users.noreply.github.com>
Co-authored-by: kkytola <“kalle.kytola@aalto.fi”>