feat: the associated sheaf of a presheaf of modules (#11555)
Let `C` be a category equipped with a Grothendieck topology `J`. Let `α : R₀ ⟶ R.val` be a morphism, with `R₀` a presheaf of rings and `R` a sheaf of rings. We assume that `α` is a sheafification map (i.e. it is both locally injective and locally surjective) . Let `M₀` be a presheaf of modules over `R₀`. Let `A` be the associated sheaf of the underlying presheaf of abelian groups of `M₀`. Then, it is shown in this PR that there is a canonical structure of sheaf of `R`-modules on `A`.



Co-authored-by: Joël Riou <37772949+joelriou@users.noreply.github.com>