feat: `refold_let` tactic (#12745)
If `x := v` is a local definition, the `refold_let x` tactic is roughly equivalent to `rw [show v = x from rfl]`. That's to say, it looks for instances of `v` and replaces them with `x`.

Adds location in formation in the `runDefEqTactic` interface so that defeq tactics can know which location is being changed, which is important for `refold_let` since `refold_let x at *` should only affect locations that come after `x`.