feat(Algebra/Homology): commutation up to signs of the compatibility isomorphisms of the total complex with shifts in both variables (#11517)
Shifting horizontally a bicomplex by an integer `x` and then taking the total complex gives a cochain complex that is isomorphic to the shift of the total complex of the original bicomplex. The same applies to vertical shifts by an integer `y`. However, when we apply both horizontal and vertical shifts, depending on whether we apply the horizontal shift or the vertical shift first, we get two isomorphisms which differ by the sign `(x * y).negOnePow`.



Co-authored-by: Joël Riou <37772949+joelriou@users.noreply.github.com>