chore: Move `CharP` lemmas about order of elements (#12816)
By swapping the order of imports between `Algebra.CharP.Basic` and `GroupTheory.OrderOfElement`, the first one goes from 944 dependencies to 812, and the second one from 912 to 915.

Also generalise a few lemmas from semirings to non-associative semirings.

This swap is justified by the fact that `CharP` is a purely algebraic concept while `OrderOf` is group-theoretic. Nevertheless,  `GroupTheory.OrderOfElement` should be split in a future PR, in which case the `CharP` lemmas might move back.